requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
python-pptx>=1.0.0
pillow>=10.0.0
